fix(auth): security + contract fixes from PR #360 review (commit 1/3)
Addresses the cross-confirmed and highest-severity findings from the
multi-agent review on PR #360.

Security (Security M3 + API m6 + Architect m6 + API C9)
- Hash Keycloak sub in backchannel-logout log line (sha256 + truncate),
  matching the codebase's hashIp pattern (GDPR Art. 4(5) — sub is a
  pseudonymous identifier, not raw-loggable at info level)
- Add 1000/min rate-limit floor on /v1/session/backchannel-logout so
  forged-token floods get capped without disrupting legitimate mass
  revocations (admin "Sign out all sessions")
- Rotate the csrf-token cookie alongside the JWT in /api/auth/refresh.
  Preserves the existing value when present (in-flight reads stay
  consistent), only mints fresh when absent. Closes the
  XSS-once-then-CSRF-forever window and prevents long-running sessions
  from silently 403ing after the original CSRF cookie expires.

Contract (API M1 + M7 + M10 + Architect M3 + M2)
- Move POST /v1/auth/backchannel-logout → POST /v1/session/backchannel-logout.
  session/service.ts already owns the blocklist primitives, sessionRoutes
  was the right module; modules/auth/ deleted entirely. OpenAPI tag is
  now Session (consistent with switch-org).
- isAuthExempt tightened from startsWith to exact pathname match so a
  future sibling route can't accidentally inherit the JWT-gate
  exemption. Path also split off query string before matching.
- Realm URL updated to /v1/session/backchannel-logout.
- Better RFC 9457 title: "Invalid logout token" (was generic "Bad
  Request"). Operators get a grep-able signal in logs.
- Wrap blocklistKeycloakSession in try/catch; Redis write failure
  returns 400 (not 500) so Keycloak surfaces the failure via its
  "Failed back-channel logout" admin event rather than mounting a
  retry storm.

Cross-package consistency (API M8 + Architect m9)
- New @givernance/shared/constants → KEYCLOAK_DEFAULT_CLIENT_ID.
  Replaces three hardcoded "givernance-web" defaults across api/env.ts
  and web/lib/auth/keycloak.ts. Env overrides still win at runtime; only
  the fallback collapses to one source of truth.
- Logout-token verifier reads env.KEYCLOAK_CLIENT_ID directly (no second
  ?? fallback). The env schema supplies the shared default, so a missing
  env shows up as an env-parse failure instead of silently defaulting.

Docs (Architect M2 + m8)
- doc 21 §4.1: add the missing jti requirement to the validation list
- doc 21 §4.1: explicit rationale for API-vs-web placement (Architect M3)
- doc 21 §4.1: new "Privacy / GDPR posture" subsection covering sid
  classification, retention, encryption-at-rest inheritance, and
  log-line policy
- doc 21 §4.1: call out scripts/keycloak-sync-realm.sh override hazard
  (operator runbook tracked as follow-up issue)
- URLs updated to /v1/session/backchannel-logout everywhere
