+++ packages/api/src/modules/session/routes.ts
@@ -1,17 +1,29 @@
 /**
- * Org-picker + switch-org routes (issue #112 / ADR-016 / doc 22 §6.3).
+ * Session-lifecycle routes (issue #112 / ADR-016 / doc 22 §6.3 + issue #76 / PR-2).
  *
  *  - `GET /v1/users/me/organizations` — cards for the picker.
- *  - `POST /v1/session/switch-org`    — validate + record + blocklist.
+ *  - `POST /v1/session/switch-org`    — validate + record + blocklist `jti`.
+ *  - `POST /v1/session/backchannel-logout` — OIDC Back-Channel Logout 1.0
+ *    webhook from Keycloak; blocklists `sid` so any in-flight access
+ *    token tied to that upstream session is rejected on next request.
  *
  * The `GET /v1/users/me` endpoint stays in the `users` module; only the
  * multi-tenant listing lives here so the `me` response is not bloated.
+ *
+ * All three handlers share the `session/service.ts` blocklist primitives
+ * (`recordOrgSwitch` writes the jti blocklist, `blocklistKeycloakSession`
+ * writes the sid blocklist). PR #360 review (API M1, M10, Architect M1)
+ * folded `modules/auth/` into this module so the contract stays coherent.
  */
 
 import { createHash } from "node:crypto";
 import { Type } from "@sinclair/typebox";
 import type { FastifyInstance, FastifyRequest } from "fastify";
 import { requireAuth } from "../../lib/guards.js";
+import {
+  type LogoutTokenClaims,
+  verifyBackchannelLogoutToken,
+} from "../../lib/keycloak-logout-token.js";
 import {
   DataArrayResponseNoPagination,
   DataResponse,
@@ -20,7 +32,12 @@ import {
   problemDetail,
   UuidSchema,
 } from "../../lib/schemas.js";
-import { listUserOrganizations, recordOrgSwitch } from "./service.js";
+import {
+  blocklistKeycloakSession,
+  KEYCLOAK_SID_BLOCKLIST_DEFAULT_TTL_S,
+  listUserOrganizations,
+  recordOrgSwitch,
+} from "./service.js";
 
 const OrgMembershipSchema = Type.Object({
   orgId: UuidSchema,
@@ -138,6 +155,137 @@ export async function sessionRoutes(app: FastifyInstance) {
       });
     },
   );
+
+  // ─── Back-channel logout (issue #76 / PR-2) ───────────────────────────
+  //
+  // OIDC Back-Channel Logout 1.0 webhook. Keycloak's `givernance-web` client
+  // is configured with `backchannel.logout.url=…/v1/session/backchannel-logout`
+  // and `backchannel.logout.session.required=true`; an upstream session end
+  // (admin "Sign out all sessions", sibling-device logout, idle timeout)
+  // triggers a server-side POST here carrying a signed `logout_token`.
+  //
+  // Verify → blocklist sid → next access-token request with that sid is
+  // rejected by the auth plugin. Exempt from JWT auth (`isAuthExempt` in
+  // `plugins/auth.ts`) — the logout token IS the authentication.
+
+  // Plugin-scoped form parser — Keycloak posts the logout token as
+  // `application/x-www-form-urlencoded`, which the global Fastify
+  // installation doesn't parse. Scoping it to this `sessionRoutes`
+  // plugin avoids accidentally enabling form parsing for unrelated
+  // v1 routes (Fastify `register` creates an encapsulation context).
+  app.addContentTypeParser(
+    "application/x-www-form-urlencoded",
+    { parseAs: "string" },
+    (_req, body, done) => {
+      const raw = typeof body === "string" ? body : body.toString("utf8");
+      const params = new URLSearchParams(raw);
+      done(null, { logout_token: params.get("logout_token") ?? "" });
+    },
+  );
+
+  /**
+   * POST /v1/session/backchannel-logout
+   *
+   * `200` on success (sid blocklisted). `400` on any verification failure
+   * — per spec, the only success/failure distinction returned to Keycloak
+   * is 200 vs 400; we never 401 or 5xx (those would prompt Keycloak to
+   * retry, which we don't want for a forged-token attempt).
+   *
+   * Rate-limit: 1000/min soft cap as a DoS floor (Security M3 / API m6).
+   * Legitimate mass-revocation bursts (admin "Sign out all sessions"
+   * across an entire tenant) stay well under this; a forged-token flood
+   * gets capped without disrupting real revocations.
+   */
+  app.post(
+    "/session/backchannel-logout",
+    {
+      config: { rateLimit: { max: 1000, timeWindow: "1 minute" } },
+      schema: {
+        tags: ["Session"],
+        body: LogoutTokenBody,
+        response: {
+          200: BackchannelLogoutResponse,
+          400: ProblemDetailSchema,
+        },
+      },
+    },
+    async (request, reply) => {
+      const { logout_token } = request.body as { logout_token: string };
+
+      let claims: LogoutTokenClaims;
+      try {
+        claims = await verifyBackchannelLogoutToken(logout_token);
+      } catch (err) {
+        request.log.warn(
+          { err: err instanceof Error ? err.message : String(err) },
+          "backchannel-logout: token verification failed",
+        );
+        return reply
+          .status(400)
+          .header("cache-control", "no-store")
+          .header("pragma", "no-cache")
+          .send(problemDetail(400, "Invalid logout token", "Logout token verification failed."));
+      }
+
+      try {
+        await blocklistKeycloakSession(claims.sid, KEYCLOAK_SID_BLOCKLIST_DEFAULT_TTL_S);
+      } catch (err) {
+        // Redis write failed. Return 400 (not 500) so Keycloak doesn't
+        // schedule a retry storm — the realm will surface the failure
+        // via the upstream "Failed back-channel logout" admin event,
+        // which is the right place for operators to find it.
+        request.log.error(
+          { err: err instanceof Error ? err.message : String(err), sid: claims.sid },
+          "backchannel-logout: Redis blocklist write failed",
+        );
+        return reply
+          .status(400)
+          .header("cache-control", "no-store")
+          .header("pragma", "no-cache")
+          .send(problemDetail(400, "Logout failed", "Could not persist session revocation."));
+      }
+
+      request.log.info(
+        {
+          sid: claims.sid,
+          // PR #360 review (API C9 — GDPR redaction posture): hash the
+          // Keycloak `sub` rather than logging it raw. `sub` is a stable
+          // pseudonymous identifier (GDPR Art. 4(5)) — pseudonymisation
+          // is fine in audit logs, but the canonical pattern in this
+          // codebase is sha256-and-truncate (see `hashIp` above).
+          subHash: claims.sub ? hashKeycloakSub(claims.sub) : undefined,
+          jti: claims.jti,
+        },
+        "backchannel-logout: sid blocklisted",
+      );
+
+      return reply
+        .status(200)
+        .header("cache-control", "no-store")
+        .header("pragma", "no-cache")
+        .send({ sid: claims.sid });
+    },
+  );
+}
+
+const LogoutTokenBody = Type.Object({
+  /**
+   * OIDC Back-Channel Logout 1.0 — § 2.4 logout token (`form-urlencoded`).
+   * Keycloak posts a single `logout_token` parameter; we verify the JWT
+   * and read `sid` out of the claims.
+   */
+  logout_token: Type.String({ minLength: 1 }),
+});
+
+const BackchannelLogoutResponse = Type.Object({
+  // Per spec the body is free-form; a tiny object with the revoked sid
+  // helps log inspection. The endpoint MUST return 200 on success and
+  // 400 on a malformed/forged token (§ 2.8).
+  sid: Type.String(),
+});
+
+function hashKeycloakSub(sub: string): string {
+  return createHash("sha256").update(sub).digest("hex").slice(0, 16);
 }
 
 type SwitchOrgError = "target_not_found" | "target_archived" | "target_suspended" | "not_a_member";
