+++ packages/shared/src/constants/index.ts
@@ -16,6 +16,7 @@ export {
   type ImpersonationEndReason,
   type ImpersonationMode,
 } from "./impersonation";
+export { KEYCLOAK_DEFAULT_CLIENT_ID } from "./keycloak";
 export { assertSafeOrgAttributes, SAFE_ORG_ATTRIBUTES } from "./keycloak-org-attributes";
 export { PASSWORD_MAX_LENGTH, PASSWORD_MIN_LENGTH } from "./password";
 export { isPersonalEmailDomain, PERSONAL_EMAIL_DOMAINS } from "./personal-email-domains";
